% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions_model.R
\name{emissions_model}
\alias{emissions_model}
\title{Simulate cumulative emissions}
\usage{
emissions_model(tsteps, A1, d, r, C)
}
\arguments{
\item{tsteps}{\code{vector} vector of time steps for integration}

\item{A1}{\code{numeric} Initial area of forest}

\item{d}{\code{numeric} Instantaneous deforestation rate}

\item{r}{\code{numeric} Instantaneous emission rate}

\item{C}{\code{numeric} Carbon stock}
}
\value{
A \code{vector} of cumulative carbon emissions.
}
\description{
Simulate cumulative emissions
}
\details{
Uses the function for deforestation in Adame et al. 2018. 
See that paper for more details
}
\examples{
years <- seq(0, 50, by = 0.1)
emissions <- emissions_model(years, 100, 0.05, 0.1, 776)
plot(years, emissions)
}
\author{
Christopher J. Brown
}
